requirements.txt
# Core ML and Computer Vision
torch>=1.8.0
torchvision>=0.9.0
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0
dlib>=19.20.0
numpy>=1.21.0
scipy>=1.7.0
scikit-image>=0.18.0

# 3D Face Reconstruction
onnxruntime>=1.8.0

# Utilities
pyyaml>=5.4.0
tqdm>=4.60.0
matplotlib>=3.3.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.0
imutils>=0.5.3
cython>=0.29.0

# Web Framework
flask>=2.0.0
flask-socketio>=5.0.0
python-socketio>=5.0.0


# Development
cmake>=3.18.0
